requirements.txt
langchain==0.3.0
langchain_community==0.3.0
langchain_core==0.3.5
langchain_groq==0.2.0
langchain_huggingface==0.1.0
langchain_text_splitters==0.3.0
python-dotenv==1.0.1
streamlit==1.38.0
pypdf==5.0.0 
